Revert "IDE-2374 use Elementals from GWT 2.5"
This reverts commit 1b7c6028e761861f17c8b867598d5d03f737d55b.

+++ exo-ide-collabeditor/src/main/java/com/google/collide/client/util/RelativeClientRect.java
@@ -37,32 +37,32 @@ public class RelativeClientRect implements ClientRect {
   }
 
   @Override
-  public float getBottom() {
+  public double getBottom() {
     return rect.getBottom() - offsetTop;
   }
 
   @Override
-  public float getHeight() {
+  public double getHeight() {
     return rect.getHeight();
   }
 
   @Override
-  public float getLeft() {
+  public double getLeft() {
     return rect.getLeft() - offsetLeft;
   }
 
   @Override
-  public float getRight() {
+  public double getRight() {
     return rect.getRight() - offsetLeft;
   }
 
   @Override
-  public float getTop() {
+  public double getTop() {
     return rect.getTop() - offsetTop;
   }
 
   @Override
-  public float getWidth() {
+  public double getWidth() {
     return rect.getWidth();
   }
 }
